refactor(subprocess_dispatch): split into focused modules (W1A)
Pure mechanical extraction of scripts/lib/subprocess_dispatch.py
(1928 lines, 22 functions) into a facade + 12 focused submodules
under scripts/lib/subprocess_dispatch_internals/.

Public API preserved — all 8 importer files (dispatch_deliver.sh,
headless_dispatch_daemon.py, claude_adapter.py, dispatch.sh,
dispatch-agent.sh, test_subprocess_canonical_receipt.py,
test_subprocess_intelligence_injection.py, test_subprocess_dispatch.py,
test_event_archival.py) work unchanged because the facade re-exports
all previously-public names (deliver_via_subprocess, deliver_with_recovery,
SubprocessAdapter, _build_intelligence_section, _inject_skill_context,
_write_manifest, _capture_dispatch_parameters, _capture_dispatch_outcome,
plus internal helpers used by tests via unittest.mock.patch).

Submodule layout:
- delivery.py / delivery_runtime.py — deliver_via_subprocess + phase helpers
- recovery.py — deliver_with_recovery + retry/success/failure branches
- skill_injection.py — _inject_skill_context, _build_intelligence_section,
  permission profile, agent cwd resolution
- manifest.py — manifest write/promote
- receipt_writer.py — _write_receipt, auto-commit, auto-stash
- pattern_confidence.py — capture_parameters/outcome, update_pattern_confidence
- git_helpers.py — commit hash, branch, dirty-files helpers
- path_utils.py — path normalization, touched-paths extraction
- handover.py — rotation handover + continuation prompt
- state_paths.py — _default_state_dir, _dispatch_manifest_dir,
  _resolve_active_dispatch_file

Function-size targets (<70L) achieved by additional Extract Method:
- deliver_via_subprocess (213L -> 60L) via _prepare_dispatch,
  _run_event_loop, _consume_event_stream, _classify_completion,
  _cleanup_dispatch_resources, _start_heartbeat
- deliver_with_recovery (158L -> 52L) via _init_recovery_state,
  _attempt_delivery, _backoff_or_fail, _handle_success,
  _handle_final_failure
- _inject_skill_context (89L) — split into discovery + content helpers
- _write_receipt (76L) — split into NDJSON-write + lease-release
- _update_pattern_confidence (89L) — split into DB-write + decision-log

Resolves: OI-1205, OI-1206, OI-1207, OI-1208, OI-1209, OI-1210
No behavior changes. No new dependencies. Tests unchanged.

Verification:
- pytest tests/test_subprocess_dispatch.py
  tests/test_subprocess_canonical_receipt.py
  tests/test_subprocess_intelligence_injection.py
  tests/test_event_archival.py -> 30 passed
- pytest tests/test_subprocess_dispatch_integration.py -> 8 passed,
  1 pre-existing failure unrelated to refactor (timeout-fail-closed
  fails on origin/main with same warning, only line number differs)
- All functions <=70 lines
- Facade <200 lines, each submodule <500 lines
- All 8 importer files parse/import cleanly

Dispatch-ID: 20260501-w1a-subprocess-dispatch-refactor
